src/backend/requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-heroku==0.3.1
django-rest-auth==0.9.3
-e git+https://github.com/wjmn/django-rest-framework-passwordless.git@master#egg=drfpasswordless
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.6.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
